# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Main Z11 api",
        "description": "This is our main Z11 api. It contains all method to handle Z11 project...",
        "termsOfService": "",
        "contact": {
            "email": "kienlv58@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/answer/edit": {
            "post": {
                "tags": [
                    "8.Answer"
                ],
                "summary": "edit answer",
                "description": "edit answer",
                "operationId": "question_edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "formData",
                        "description": "question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "answer",
                        "in": "formData",
                        "description": "answer type json",
                        "required": true,
                        "schema": {
                            "required": [
                                "answer"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
        "/group_question/get/{id}": {
            "get": {
                "tags": [
                    "6.GroupQuestion"
                ],
                "summary": "get group_question",
                "description": "get group_question with group_question_id",
                "operationId": "get group_question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "group_question_id",
                        "in": "formData",
                        "description": "group_question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "grant_type"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get succes"
                    },
                    "400": {
                        "description": "Invalid param"
                    }
                }
            }
        },
        "/group_question/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "6.GroupQuestion"
                ],
                "summary": "get all group_question",
                "description": "return group_question with take and skip",
                "operationId": "group_question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
...
        "/question/get/{id}": {
            "get": {
                "tags": [
                    "7.Question"
                ],
                "summary": "get question",
                "description": "get group_question with question_id",
                "operationId": "get question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "formData",
                        "description": "question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "grant_type"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get succes"
                    },
                    "400": {
                        "description": "Invalid param"
                    }
                }
            }
        },
        "/question/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "7.Question"
                ],
                "summary": "get all question",
                "description": "return question with take and skip",
                "operationId": "question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
        "/question/add": {
            "post": {
                "tags": [
                    "7.Question"
                ],
                "summary": "add new question",
                "description": "add new question",
                "operationId": "question_add",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "group_question_id",
                        "in": "formData",
                        "description": "group_question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "group_question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "sub_question_content",
                        "in": "formData",
                        "description": "sub_question_content",
                        "required": true,
                        "schema": {
                            "required": [
                                "sub_question_content"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "translate",
                        "in": "formData",
                        "description": "translate json",
                        "required": true,
                        "schema": {
                            "required": [
                                "category_code"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "explain_cost",
                        "in": "formData",
                        "description": "explain_cost",
                        "required": true,
                        "schema": {
                            "required": [
                                "explain_cost"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
        "/question/edit": {
            "post": {
                "tags": [
                    "7.Question"
                ],
                "summary": "edit group_question",
                "description": "edit question",
                "operationId": "question_edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "formData",
                        "description": "question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "sub_question_content",
                        "in": "formData",
                        "description": "sub_question_content",
                        "required": true,
                        "schema": {
                            "required": [
                                "sub_question_content"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "translate",
                        "in": "formData",
                        "description": "translate json",
                        "required": true,
                        "schema": {
                            "required": [
                                "category_code"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "explain_cost",
                        "in": "formData",
                        "description": "explain_cost",
                        "required": true,
                        "schema": {
                            "required": [
                                "explain_cost"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
        "/question/delete": {
            "post": {
                "tags": [
                    "7.Question"
                ],
                "summary": "delete group_question ",
                "description": "delete with question",
                "operationId": "questiondelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "formData",
                        "description": "uid delete",
                        "required": true,
                        "schema": {
                            "required": [
                                "uid"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "question_id",
                        "in": "formData",
                        "description": "question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
    },
...
}